new name: orch_db
Replace the temporary 'squeel' name. It was a placeholder until I could
decide on the general name used for the set of cookbooks that work
together.

Orch is short for "Orchestrate". The various orch cookbooks are used to
'orchestrate' the provisioning of one or more servers for your
rails/rack app.